// projects/sl-questionnaire/src/lib/interfaces/questionnaire.type.ts
export type Code = string;
// ...
export interface Question {
  _id: string;
  question: string;
  isCompleted: boolean;
  showRemarks: string;
  options: string;
  option?:{hint:"",tip:""}[]
  sliderOptions: string;
  children: string[];
  questionGroup: string;
  fileName: string;
  instanceQuestions: string;
  isAGeneralQuestion: string;
  autoCapture: string;
  allowAudioRecording: string;
  prefillFromEntityProfile: string;
  entityFieldName: string;
  isEditable: string;
  showQuestionInPreview: string;
  deleted: string;
  remarks: string;
  value: string | string[];
  usedForScoring: string;
  questionType: string;
  canBeNotApplicable: string;
  visibleIf: VisibleIfUnion;
  validation: string | Validation;
  dateFormat: string;
  externalId: string;
  tip: string;
  hint: string;
  responseType: ResponseType;
  modeOfCollection: string;
  accessibility: string;
  rubricLevel: string;
  sectionHeader: string;
  page: string;
  questionNumber: string;
  updatedAt: string;
  createdAt: string;
  __v: string;
  createdFromQuestionId: string;
  evidenceMethod: string;
  payload: Payload;
  startTime: number;
  endTime: number;
  gpsLocation: string;
  file: string;
  pageQuestions: Question[];
}
// ...
export enum ResponseType {
  TEXT = 'text',
  NUMBER = 'number',
  RADIO = 'radio',
  MULTISELECT = 'multiselect',
  DATE = 'date',
  SLIDER = 'slider',
  PAGEQUESTIONS = 'pageQuestions',
  MATRIX = 'matrix',
}

export type FileUnion = FileClass | string;

export interface FileClass {
  required: boolean;
  type: string[];
  minCount: number;
  maxCount: number;
  caption: string;
}

export interface Payload {
  criteriaId: string;
  responseType: string;
  evidenceMethod: Code;
  rubricLevel: string;
}

export interface Validation {
  required: boolean;
  max?: string;
  min?: string;
  IsNumber?: string;
  regex: RegExp;
}

export interface Option {
  value: string;
  label: string;
}

export type VisibleIfUnion = VisibleIfElement[] | string;

export interface VisibleIfElement {
  operator: string;
  value: string[];
  _id: string;
}